requirements.txt
django==2.1.5
pillow==5.4.1
django-ckeditor==5.6.1
django-import-export==1.2.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
celery==4.1.1
rabbitmq==0.2.0
pytils==0.3
django-jet==1.0.8
